zaed_sql_agent-main/requirements.txt
langchain
langchain-core
langchain-experimental
langchain-openai
langchain-mcp-adapters
fastmcp
sqlalchemy
psycopg2-binary
python-dotenv
langchain-community